refactor(engine): split libretro_engine_napi into modular components for improved maintainability and clarity
